Refactor: 상품 상세페이지 통합
Refs: #7
