+++ src/App.js
@@ -5,7 +5,7 @@ import JoinPage from './pages/joinPage/JoinPage';
 import CartPage from './pages/cartPage/CartPage';
 import PaymentPage from './pages/paymentPage/PaymentPage';
 import NotFound from './pages/notFoundPage/NotFoundPage';
-import ProductDetailPage from './pages/ProductDetailPage';
+import ProductDetailPage from './pages/productDetailPage/ProductDetailPage';
 import SellerCenterPage from './pages/SellerCenterPage';
 import EditProductPage from './pages/editProductPage/EditProductPage';
 import OrderCompletedPage from './pages/orderCompletedPage/OrderCompletedPage';
